Split out common layout integration into separate layout component
This new component is called LayoutBase (as opposed to BaseLayout)
because unlike the other layouts it cannot be used directly
from Markdown (because of the required "title" prop). Still, it is kept
in the layouts directory in order to ensure that everything in the
"components" does not require the integration.